refactor switch to if and correct naming of api route to movies
